refactor: move algorithm implementation to -impl.h file
